Move examples into the library to be able to instance them
